requirements.txt
Flask==2.3.4
tensorflow==2.14.0
numpy==1.26.5
Pillow==10.1.0
gunicorn==23.2.0
